Move tusCtrl to common controllers
